session 정보 불러오는 로직 변경
util패키지의 SessionUtils에서 세션 정보를 받아 오는 것으로 바꿈

+++ src/main/java/com/likelion12_team_project/controller/GonggamCommentController.java
@@ -4,13 +4,14 @@ import com.likelion12_team_project.dto.request.GonggamCommentRequest;
 import com.likelion12_team_project.dto.response.GonggamCommentResponse;
 import com.likelion12_team_project.entity.User;
 import com.likelion12_team_project.service.GonggamCommentService;
+import com.likelion12_team_project.util.SessionUtils;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 
 @RestController
 @RequestMapping("/api/gonggamcomments")
@@ -22,24 +23,39 @@ public class GonggamCommentController {
 
     @PostMapping("/{postId}")
     public ResponseEntity<GonggamCommentResponse> addComment(@PathVariable("postId") Long postId, @RequestBody GonggamCommentRequest commentRequest, HttpServletRequest request) {
-        HttpSession session = request.getSession(false);
-        User user = (User) session.getAttribute("user");
+
+    	ResponseEntity<User> userResponse = SessionUtils.getCurrentUser(request);
+    	if (userResponse.getStatusCode() != HttpStatus.OK) {
+    	    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+    	}
+    	User user = userResponse.getBody();
+    	
         GonggamCommentResponse comment = gonggamCommentService.addComment(commentRequest, user.getId(), postId);
         return new ResponseEntity<>(comment, HttpStatus.CREATED);
     }
 
     @PutMapping("/{postId}/{commentId}")
     public ResponseEntity<GonggamCommentResponse> updateComment(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId, @RequestBody GonggamCommentRequest commentRequest, HttpServletRequest request) {
-        HttpSession session = request.getSession(false);
-        User user = (User) session.getAttribute("user");
+
+    	ResponseEntity<User> userResponse = SessionUtils.getCurrentUser(request);
+    	if (userResponse.getStatusCode() != HttpStatus.OK) {
+    	    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+    	}
+    	User user = userResponse.getBody();
+    	
         GonggamCommentResponse comment = gonggamCommentService.updateComment(commentId, commentRequest, user.getId());
         return ResponseEntity.ok(comment);
     }
 
     @DeleteMapping("/{postId}/{commentId}")
     public ResponseEntity<Void> deleteComment(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId, HttpServletRequest request) {
-        HttpSession session = request.getSession(false);
-        User user = (User) session.getAttribute("user");
+        
+    	ResponseEntity<User> userResponse = SessionUtils.getCurrentUser(request);
+    	if (userResponse.getStatusCode() != HttpStatus.OK) {
+    	    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+    	}
+    	User user = userResponse.getBody();
+
         gonggamCommentService.deleteComment(commentId, user.getId());
         return ResponseEntity.noContent().build();
     }
